# frontend/client-code-gen/fam-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Forest Access Management - FAM - API",
    "description" : "Forest Access Management API. Defines who has access to which applications, and the roles they operate under once access is granted.\n\nCovers both the app-access-control and admin-management surfaces that upstream FAM served from two separate APIs.",
    "contact" : {
      "name" : "Team Heartwood",
      "url" : "https://apps.nrs.gov.bc.ca/int/confluence/display/FSAST1/Team+Heartwood",
      "email" : "SIBIFSAF@Victoria1.gov.bc.ca"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this document"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/application-admins" : {
      "get" : {
        "tags" : [ "FAM Application Admin" ],
        "summary" : "All application admins",
        "operationId" : "get_application_admins",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FamAppAdminGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "FAM Application Admin" ],
        "summary" : "Make a user an administrator of an application",
        "operationId" : "create_application_admin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FamAppAdminCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamAppAdminGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/access-control-privileges" : {
      "get" : {
        "tags" : [ "FAM Access Control Privileges" ],
        "summary" : "Delegated admin privileges for an application",
        "operationId" : "get_access_control_privileges_by_application_id",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "description" : "Page number",
            "minimum" : 1
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Number of records per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50,
            "description" : "Number of records per page",
            "maximum" : 100000,
            "minimum" : 10
          }
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Search by keyword",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Search by keyword",
            "maxLength" : 30,
            "minLength" : 3
          }
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "description" : "Column sorting order",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortOrder"
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "description" : "Column to be sorted by",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UserRoleSortBy"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResultsFamAccessControlPrivilegeGetResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "FAM Access Control Privileges" ],
        "summary" : "Grant delegated admin privileges",
        "operationId" : "create_access_control_privilege_many",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FamAccessControlPrivilegeCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamAccessControlPrivilegeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/application-admins/application/{applicationId}" : {
      "get" : {
        "tags" : [ "FAM Application Admin" ],
        "summary" : "Application admins for one application",
        "operationId" : "get_application_admins_by_application_id",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FamAppAdminGetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/access-control-privileges/export" : {
      "get" : {
        "tags" : [ "FAM Access Control Privileges" ],
        "summary" : "Export delegated admin roles by application ID",
        "operationId" : "export_access_control_privileges_by_application_id",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/application-admins/{applicationAdminId}" : {
      "delete" : {
        "tags" : [ "FAM Application Admin" ],
        "summary" : "Remove an application administrator",
        "operationId" : "delete_application_admin",
        "parameters" : [ {
          "name" : "applicationAdminId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/access-control-privileges/{accessControlPrivilegeId}" : {
      "delete" : {
        "tags" : [ "FAM Access Control Privileges" ],
        "summary" : "Remove a delegated admin privilege",
        "operationId" : "delete_access_control_privilege",
        "parameters" : [ {
          "name" : "accessControlPrivilegeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "UserType" : {
        "type" : "string",
        "enum" : [ "I", "B", "CD", "CT", "CP" ]
      },
      "EmailSendingStatus" : {
        "type" : "string",
        "enum" : [ "NOT_REQUIRED", "SENT_TO_EMAIL_SERVICE_SUCCESS", "SENT_TO_EMAIL_SERVICE_FAILURE" ]
      },
      "FamApplicationDto" : {
        "properties" : {
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "application_name" : {
            "type" : "string"
          },
          "application_description" : {
            "type" : "string"
          }
        },
        "required" : [ "application_description", "application_id", "application_name" ]
      },
...
      "FamForestClientDto" : {
        "properties" : {
          "client_name" : {
            "type" : "string"
          },
          "forest_client_number" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/FamForestClientStatusDto"
          }
        },
        "required" : [ "forest_client_number" ]
      },
      "FamForestClientStatusDto" : {
        "properties" : {
          "status_code" : {
            "$ref" : "#/components/schemas/FamForestClientStatusType"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "FamForestClientStatusType" : {
        "type" : "string",
        "enum" : [ "A", "I" ]
      },
      "FamRoleMinDto" : {
        "properties" : {
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          }
        },
        "required" : [ "application", "role_name", "role_type_code" ]
      },
      "FamRoleWithClientDto" : {
        "properties" : {
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "display_name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          },
          "forest_client" : {
            "$ref" : "#/components/schemas/FamForestClientDto"
          },
          "parent_role" : {
            "$ref" : "#/components/schemas/FamRoleMinDto"
          }
        },
        "required" : [ "application", "description", "forest_client", "role_id", "role_name", "role_type_code" ]
      },
      "FamUserInfoDto" : {
        "properties" : {
          "user_name" : {
            "type" : "string"
          },
          "user_type" : {
            "$ref" : "#/components/schemas/FamUserTypeDto"
          },
          "first_name" : {
            "type" : "string"
          },
          "last_name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "user_name", "user_type" ]
      },
...
      "FamUserTypeDto" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "description" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "description" ]
      },
...
      "FamAppAdminCreateRequest" : {
        "properties" : {
          "user_name" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 3
          },
          "user_guid" : {
            "type" : "string",
            "maxLength" : 32,
            "minLength" : 32
          },
          "user_type_code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "application_id", "user_guid", "user_name", "user_type_code" ]
      },
      "AppEnv" : {
        "type" : "string",
        "enum" : [ "DEV", "TEST", "PROD" ]
      },
      "FamAppAdminGetResponse" : {
        "properties" : {
          "application_admin_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "create_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "user" : {
            "$ref" : "#/components/schemas/FamUserInfoDto"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationBase"
          }
        },
        "required" : [ "application", "application_admin_id", "application_id", "user", "user_id" ]
      },
      "FamApplicationBase" : {
        "properties" : {
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "application_name" : {
            "type" : "string"
          },
          "application_description" : {
            "type" : "string"
          },
          "app_environment" : {
            "$ref" : "#/components/schemas/AppEnv"
          }
        },
        "required" : [ "application_description", "application_id", "application_name" ]
      },
      "FamAccessControlPrivilegeCreateRequest" : {
        "properties" : {
          "user_name" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 3
          },
          "user_guid" : {
            "type" : "string",
            "maxLength" : 32,
            "minLength" : 32
          },
          "user_type_code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "forest_client_numbers" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 8,
              "minLength" : 1
            }
          },
          "requires_send_user_email" : {
            "type" : "boolean"
          }
        },
        "required" : [ "role_id", "user_guid", "user_name", "user_type_code" ]
      },
      "FamAccessControlPrivilegeCreateResponse" : {
        "properties" : {
          "status_code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "$ref" : "#/components/schemas/FamAccessControlPrivilegeGetResponse"
          },
          "error_message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "required" : [ "detail", "status_code" ]
      },
      "FamAccessControlPrivilegeGetResponse" : {
        "properties" : {
          "access_control_privilege_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/FamUserInfoDto"
          },
          "role" : {
            "$ref" : "#/components/schemas/FamRoleWithClientDto"
          },
          "create_date" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "access_control_privilege_id", "create_date", "role", "role_id", "user", "user_id" ]
      },
      "FamAccessControlPrivilegeResponse" : {
        "properties" : {
          "email_sending_status" : {
            "$ref" : "#/components/schemas/EmailSendingStatus"
          },
          "assignments_detail" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamAccessControlPrivilegeCreateResponse"
            }
          }
        },
        "required" : [ "assignments_detail" ]
      },
...
      "SortOrder" : {
        "type" : "string",
        "description" : "Column sorting order",
        "enum" : [ "asc", "desc" ]
      },
      "UserRoleSortBy" : {
        "type" : "string",
        "description" : "Column to be sorted by",
        "enum" : [ "create_date", "user_name", "user_type_code", "email", "full_name", "role_display_name", "forest_client_number" ]
      },
      "PageResultMeta" : {
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total records counts for query conditions"
          },
          "number_of_pages" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Total pages for query records"
          },
          "page_number" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page number"
          },
          "page_size" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of records per page"
          }
        },
        "required" : [ "number_of_pages", "page_number", "page_size", "total" ]
      },
...
      "PagedResultsFamAccessControlPrivilegeGetResponse" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/PageResultMeta"
          },
          "results" : {
            "type" : "array",
            "description" : "Paged results",
            "items" : {
              "$ref" : "#/components/schemas/FamAccessControlPrivilegeGetResponse"
            }
          }
        },
        "required" : [ "meta", "results" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Access token issued by BC Gov SSO (Keycloak).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "Shared secret for the scheduled user-information refresh.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}